requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
cffi==1.17.1
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.18.3
yt-dlp==2025.1.26